src/brain/pyproject.toml
[project]
name = "brain"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "redis",
    "psycopg2-binary",
    "pydantic-settings",
]